requirements.txt
# Core OCR and Computer Vision
pytesseract==0.3.10
paddleocr==2.7.3
easyocr==1.7.0
opencv-python==4.8.1.78
Pillow==10.0.1
pdf2image==1.16.3

# NLP and Text Processing
spacy==3.7.2
nltk==3.8.1
transformers==4.35.2
sentence-transformers==2.2.2

# LLM APIs
openai==1.3.8
google-generativeai==0.3.2
anthropic==0.7.8

# Vector and SQL Databases
milvus-lite==2.3.3
chromadb==0.4.15
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
sqlite3

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
chainlit==1.0.200
streamlit==1.28.1
pydantic==2.5.0

# ML and Evaluation
scikit-learn==1.3.2
wandb==0.16.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.25.2

# Additional utilities
python-dotenv==1.0.0
requests==2.31.0
asyncio
aiofiles==23.2.1
python-multipart==0.0.6

# Text similarity and evaluation
rouge-score==0.1.2
python-Levenshtein==0.23.0
bert-score==0.3.13

# Image processing
imageio==2.33.1
wand==0.6.13

# Development tools
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1